* [MOD] Code refactoring
